+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.0.1
 
 * Android native battery level integrated
-* MIT & Apache licence added
+* MIT & Apache License added
+
+## 0.1.1
+* Added support for retrieving native device battery level on iOS.
+* Implemented platform communication using Flutter Method Channels.